tasks/Sber/ml_solution/requirements.txt
# Основные зависимости
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.2.0+cu118
hydra-core==1.3.2
omegaconf==2.3.0
numpy==1.24.3
pandas==2.1.4
networkx==3.2.1
tqdm==4.66.1

# ML и оптимизация
scikit-learn==1.3.2
optuna==3.4.0
mlflow==2.9.2

# Логирование и визуализация
tensorboard==2.15.1
matplotlib==3.8.2
seaborn==0.13.1

# Утилиты
pyyaml==6.0.1

# Линтинг и форматирование
ruff==0.2.1
mypy==1.8.0

# Тестирование
pytest==7.4.3
pytest-cov==4.1.0 